0-over-lim (#8)
Based on [this conversation](https://leanprover.zulipchat.com/#narrow/channel/487278-toric/topic/0-over-lim.3A.20Limit-preserving.20functors.20lift).

There are two theorems:
- If a functor `F` preserves limits of one shape, `Over.post F` preserves limits of a different shape (below)
- If a functor `F` preserves all finite limits, so does `Over.post F`

The different shape is constructed by adding another object to a category and making it the final object. It's the same as `WidePullbackShape`, except that this one only allows you to extend a **discrete** category.