Diagonalisable Hopf algebras (#17)
This introduces diagonalisable bialgebras as the bialgebras that are isomorphic to the monoid algebra of a (not necessarily commutative) group. Then it proves that a bialgebra over a field is diagonalisable if and only if ti is generated by its group-like elements.

It also repairs the statement of `Scheme.isDiagonalisable_iff_span_isGroupLikeElem_eq_top` by adding the `IsAffine` condition.