Foundations for an algebraic theory of polyhedral cones (#23)
This PR ports existing results for polyhedral cones into a completely "algebraic" setting. Everything is stated for a pair of modules over an arbitrary linearly ordered field with a bilinear pairing (as opposed to working over the real numbers). The modules should both be finitely generated and the bilinear pairing injective in both arguments for everything to work nicely. This should be a suitable setting for the purposes of toric geometry.

The fundamental result here is that polyhedral cones (defined to be the intersection of finitely many halfspaces, i.e. the dual of a finite set) are precisely the finitely generated ones. Results like that the dual of a polyhedral cone is again polyhedral or that the double dual of a finitely generated cone is the cone itself follow easily from that. Obvious next steps would be to define faces of cones, relative interiors, the category of lattice cones and lattice fans. Everything should be doable in this algebraic setting, no topology should be required.

I've copied the contents of the file `Mathlib/Geometry/Convex/Dual.lean` from the currently open leanprover-community/mathlib4#24149, which provides some basic API around the algebraic dual of a cone. Besides this, there is not much overlap with that PR.

Co-authored-by: Yaël Dillies <yael.dillies@gmail.com>